refactor: token 패키지를 auth 패키지로 변경
