+++ src/main/java/com/depromeet/domain/auth/dao/RefreshTokenRepository.java
@@ -1,6 +1,7 @@
-package com.depromeet.domain.token.dao;
+package com.depromeet.domain.auth.dao;
 
-import com.depromeet.domain.token.domain.RefreshToken;
+import com.depromeet.domain.auth.domain.RefreshToken;
+
 import org.springframework.data.repository.CrudRepository;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {}
