+++ src/main/java/com/depromeet/domain/auth/domain/RefreshToken.java
@@ -1,7 +1,8 @@
-package com.depromeet.domain.token.domain;
+package com.depromeet.domain.auth.domain;
 
 import lombok.Builder;
 import lombok.Getter;
+
 import org.springframework.data.annotation.Id;
 import org.springframework.data.redis.core.RedisHash;
 import org.springframework.data.redis.core.TimeToLive;
