+++ src/main/java/com/depromeet/global/security/JwtTokenProvider.java
@@ -2,18 +2,22 @@ package com.depromeet.global.security;
 
 import static com.depromeet.global.common.constants.SecurityConstants.TOKEN_ROLE_NAME;
 
+import com.depromeet.domain.auth.dto.AccessToken;
 import com.depromeet.domain.member.domain.MemberRole;
-import com.depromeet.domain.token.dto.AccessToken;
 import com.depromeet.global.error.exception.CustomException;
 import com.depromeet.global.error.exception.ErrorCode;
 import com.depromeet.infra.config.jwt.JwtProperties;
+
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
-import java.security.Key;
-import java.util.Date;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.stereotype.Component;
+
+import java.security.Key;
+import java.util.Date;
 
 @Slf4j
 @Component
